requirements.txt
python-telegram-bot==20.7
stellar-sdk==13.0.0
python-dotenv==1.0.0
aiohttp==3.9.1
pymongo==4.6.1
